TOTS-22 Moved the getLayerName function to the shared utils file.
