+++ app/client/src/utils/utils.ts
@@ -1,3 +1,5 @@
+import { LayerType } from 'types/Layer';
+
 /**
  * Script from ESRI for escaping an ArcGIS Online usernames and
  * organization ids.
@@ -79,4 +81,27 @@ export function createErrorObject(error: any) {
  */
 export function escapeRegex(str: string) {
   return str.replace(/([.*+?^=!:${}()|\]\\])/g, '\\$1');
+}
+
+/**
+ * Determines if the desired name has already been used. If it has
+ * it appends in index to the end (i.e. '<desiredName> (2)').
+ */
+export function getLayerName(layers: LayerType[], desiredName: string) {
+  // get a list of names in use
+  let usedNames: string[] = [];
+  layers.forEach((layer) => {
+    usedNames.push(layer.label);
+  });
+
+  // Find a name where there is not a collision.
+  // Most of the time this loop will be skipped.
+  let duplicateCount = 0;
+  let newName = desiredName;
+  while (usedNames.includes(newName)) {
+    duplicateCount += 1;
+    newName = `${desiredName} (${duplicateCount})`;
+  }
+
+  return newName;
 }
